fix(kernel,extensions): unbreak workspace build
Two pre-existing breakages on main, dragged in by separate merges
that didn't actually compile against each other:

1. resolve_cron_max_tokens took Option<u32> but cron_session_max_tokens
   in KernelConfig is Option<u64> — call site at kernel/mod.rs:13335
   failed to compile. Widen the helper to u64 (it's just zero-checking,
   not arithmetic).

2. librefang-extensions exposed http_client as pub(crate), but
   librefang-kernel::mcp_oauth_provider calls
   librefang_extensions::http_client::new_client() across crates.
   Promote the module to pub.

+++ crates/librefang-kernel/src/kernel/mod.rs
@@ -94,7 +94,7 @@ pub(crate) fn resolve_cron_max_messages(raw: Option<usize>) -> Option<usize> {
 /// - `None`    → no cap
 /// - `Some(0)` → disable (treat as no cap)
 /// - `Some(n)` otherwise → use as-is
-pub(crate) fn resolve_cron_max_tokens(raw: Option<u32>) -> Option<u32> {
+pub(crate) fn resolve_cron_max_tokens(raw: Option<u64>) -> Option<u64> {
     match raw {
         Some(0) => None,
         other => other,
